validation-framework/requirements.txt
pyyaml==6.0.2
pytest==8.3.2
flask